+++ src/test/testUtils.tsx
@@ -17,20 +17,15 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-module.exports = {
-  preset: 'ts-jest',
-  testEnvironment: 'jest-environment-jsdom',
-  testMatch: ['**/*.test.{ts,tsx,js,jsx}'],
-  transform: {
-    '^.+\\.tsx?$': 'ts-jest',
-  },
-  moduleNameMapper: {
-    '\\.(css|scss)$': '<rootDir>/src/test/__mocks__/styleMock.ts',
-    '\\.svg$': '<rootDir>/src/test/__mocks__/fileMock.ts',
-    '@mui/material/styles/createPalette':
-      '<rootDir>/src/test/__mocks__/createPalette.ts',
-    '@mui/material/styles/createTypography':
-      '<rootDir>/src/test/__mocks__/createTypography.ts',
-  },
-  setupFiles: ['<rootDir>/src/test/setupTests.ts'],
+import React, { type ReactElement } from 'react'
+import { SharedThemeProvider } from '../components'
+import { render as rtlRender, type RenderOptions } from '@testing-library/react'
+
+const AllTheProviders = ({ children }: { children: React.ReactNode }) => {
+  return <SharedThemeProvider>{children}</SharedThemeProvider>
 }
+
+export const render = (
+  ui: ReactElement,
+  options?: Omit<RenderOptions, 'wrapper'>
+) => rtlRender(ui, { wrapper: AllTheProviders, ...options })